+++ src/main/java/cz/morosystems/worklogger/common/WorklogService.java
@@ -1,25 +1,11 @@
 package cz.morosystems.worklogger.common;
 
-import com.atlassian.jira.rest.client.api.JiraRestClient;
-import com.atlassian.jira.rest.client.api.SearchRestClient;
 import com.atlassian.jira.rest.client.api.domain.Issue;
-import com.atlassian.jira.rest.client.api.domain.SearchResult;
-import com.atlassian.jira.rest.client.internal.async.AsynchronousJiraRestClientFactory;
-import com.atlassian.util.concurrent.Promise;
-import java.io.IOException;
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.HashMap;
 import org.json.JSONObject;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-/**
- * Created by moro on 9/27/2017.
- */
-public class WorklogManipulation {
+public class WorklogService {
   private static final DateTimeFormatter dateTimeFormat = DateTimeFormatter
       .ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSSZ");
 
@@ -54,9 +40,5 @@ public class WorklogManipulation {
 
   }
 
-
-  /**
-   * Created by moro on 9/20/2017.
-   */
 
 }
